allow importing another app; add redis dump to ignore
